updated html code and added 2 new css files
removed blue-box and two links from the top section of the html and then added in an About Me section. Also added in a haverable navigation links to the top section.

added the code for the hoverable navigation into the hoverableSidebar.css file

created an aboutMe.css file as well