updated html and css files
updated index.html to put the img in top-section on the left and info on the right. added in a div with the class=bounce to add an animation on the img.

removed blue-square and links in the top-section.

updated the height and width of the img to be the same and changed the border-radius to be 50% to make it a circle.

updated top-section.css to include code for the animation of bounce to the img.

updated font color for the experience.css, footer.css (hover color), top-section.css & portfolio.css sections to a more red color using hsl().

updated background in portfolio.css file to a lighter gray color using hsl().